requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.3.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cloudpickle==1.2.2
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
dill==0.3.1.1
dm-control==0.0.288483845
dm-env==1.2
dm-tree==0.1.1
easydict==1.9
fasteners==0.15
future==0.18.2
gast==0.2.2
glfw==1.10.1
google-pasta==0.1.7
googleapis-common-protos==1.6.0
grpcio==1.23.0
h5py==2.9.0
idna==2.8
ignite==1.1.0
imageio==2.6.1
joblib==0.14.0
jsonpatch==1.26
jsonpointer==2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-rectified-adam==0.17.0
kiwisolver==1.1.0
lmdb==0.98
lxml==4.4.2
Markdown==3.1.1
matplotlib==3.1.1
monotonic==1.5
mujoco-py==2.0.2.9
networkx==2.4
numpy==1.16.2
opencv-python==4.1.0.25
Pillow==6.1.0
promise==2.2.1
protobuf==3.9.1
pycparser==2.19
pyglet==1.3.2
PyOpenGL==3.1.5
pyparsing==2.4.2
python-dateutil==2.8.0
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==19.0.2
requests==2.22.0
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
style==1.1.0
tensorboard==1.14.0
tensorboardX==2.1
tensorflow-datasets==1.3.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-metadata==0.15.1
tensorflow-probability==0.7.0
termcolor==1.1.0
torch==1.2.0
torchfile==0.1.0
torchvision==0.4.0
tornado==6.0.4
tqdm==4.38.0
transforms3d==0.3.1
update==0.0.1
urllib3==1.25.7
visdom==0.1.8.9
websocket-client==0.57.0
Werkzeug==0.15.5
wrapt==1.11.2
xgboost==0.90
